+++ HISTORY.md
@@ -2,6 +2,7 @@ Next Release
 ------
  * Added check to see if Phantom.js is installed
  * Added .travis.yml for unit testing on Travis CI
+ * Added `--loglevel`/`--debug` support to yb-compile
 
 v0.1.0
 ------